feat: pass in Options Object to http Constructor (#37)
It looks like the function was written to allow this as a parameter, but it was never used. This should allow for use cases like specifying custom headers while using `urlSource`. The `node-fetch` module falls back on [default options](https://github.com/node-fetch/node-fetch#options) if none are specified, so we shouldn't need to check and send an empty object at this step.