fix: make timeouts stricter
Sometimes a request can resolve before the timeout callback fires,
this can be because the system is under load and the event loop is
blocked.

This change looks at the time taken by a request to resolve when
deciding if it's timed out as well as setting a regular js timeout
to catch inactive requests.

It also removes the `timeout` option passed to `node-fetch` as it
implements it's own timeout feature which results in two types of
timeout error that can occur - ours and `node-fetch`'s.