fix: use native fetch if available
If you are in the Electron Renderer process, native fetch is available
but it's ignored because the renderer does not respect the browser
field of package.json, so we end up with node-fetch that uses a
browser polyfill of the node http api.

Instead use native fetch if it's available or node fetch if not.