fix: use node-fetch fork with fix for stream closed prematurely bug
https://github.com/node-fetch/node-fetch/pull/1064 fixes a bug in node-fetch
to make it handle situations where the stream closes prematurely but
it's been merged into the v3 release tree which is still future tech
with no release date.

https://github.com/node-fetch/node-fetch/pull/1172 backports that fix
to v2 but although approved it's not been merged and released yet
so here we use a temporary fork published with that PR merged in.